feat: refactor review submission logic; replace CustomerReview component with CustomerReviewModal and enhance order item rating handling
