+++ src/app/api/purchase/rate/route.ts
@@ -7,8 +7,7 @@ export async function POST(req: NextRequest) {
     const body = await req.json().catch(() => ({}));
     const { product_id, rating, review_text, order_id } = body;
 
-
-    console.log("Received review submission:", { product_id, rating, review_text, order_id });
+    
 
     // basic validation
     if (!product_id) {
@@ -20,15 +19,15 @@ export async function POST(req: NextRequest) {
     }
 
     const userId = await getUserId(req);
-    
+
     if (!userId) {
       return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
     }
 
     const payload = {
       user_id: userId,
-      product_id,
-      order_id,
+      product_id: Number(product_id),
+      order_id: order_id ?? null,
       rating: parsedRating,
       review_text: review_text ?? null,
       created_at: new Date().toISOString(),
@@ -40,8 +39,30 @@ export async function POST(req: NextRequest) {
       return NextResponse.json({ error: error.message }, { status: 400 });
     }
 
-    return NextResponse.json({ success: true, review: data }, { status: 201 });
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    // Flip is_rated to true on matching order_items row(s)
+    let updateResult = null;
+    try {
+      const matchObj: Record<string, unknown> = { product_id: payload.product_id };
+      // only include order_id in the match if provided (not null/undefined)
+      if (payload.order_id != null) matchObj.order_id = payload.order_id;
+
+      const { data: updateData, error: updateError } = await supabaseServer
+        .from("order_items")
+        .update({ is_rated: true })
+        .match(matchObj);
+
+      if (updateError) {
+        // log but don't fail the whole request
+        console.error("Failed to update order_items.is_rated:", updateError);
+      } else {
+        updateResult = updateData;
+      }
+    } catch (err) {
+      console.error("Error updating order_items.is_rated:", err);
+    }
+
+    return NextResponse.json({ success: true, review: data, updatedOrderItems: updateResult }, { status: 201 });
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
   } catch (err: any) {
     return NextResponse.json({ error: err?.message || "Server error" }, { status: 500 });
   }
