+++ src/components/user/purchase/ItemList.tsx
@@ -31,8 +31,6 @@ const ItemList: React.FC = () => {
     setLoading(false);
   };
 
-  console.log("ItemList purchases:", purchases);
-
   const fetchStatusCounts = async () => {
     const statuses = Object.keys(statusMap).filter((s) => s !== "All");
     const counts: Record<string, number> = {};
